🚧 ✅ 🔧  Refactor and turn off Playwright console logging to log-utils.ts, and turn back on full test suite.
